LoggedIn/LoggedOut components factored into separate files
