+++ client/src/components/Navbar/LoggedOut/useStyles.ts
@@ -1,0 +1,7 @@
+import { makeStyles } from '@material-ui/core/styles';
+
+const useStyles = makeStyles(() => ({
+  button: { marginRight: '0.5rem' },
+}));
+
+export default useStyles;
