+++ client/src/components/Navbar/useStyles.ts
@@ -1,6 +1,6 @@
-import { makeStyles, Theme } from '@material-ui/core/styles';
+import { makeStyles } from '@material-ui/core/styles';
 
-const useStyles = makeStyles((theme: Theme) => ({
+const useStyles = makeStyles(() => ({
   toolbar: { minHeight: 80 },
   logo: {
     flexGrow: 1,
@@ -16,10 +16,6 @@ const useStyles = makeStyles((theme: Theme) => ({
     marginRight: 30,
   },
   button: { marginRight: '0.5rem' },
-  avatar: {
-    width: theme.spacing(6),
-    height: theme.spacing(6),
-  },
 }));
 
 export default useStyles;
